requirements.txt
streamlit
pandas
Pillow
numpy
tensorflow
opencv-python
